chore(utils): cleanup
