--- utils/deepCopy/index.js
@@ -1,49 +1,0 @@
-/**
- * Get the first item that pass the test
- * by second argument function
- *
- * @param {Array} list
- * @param {Function} f
- * @return {*}
- */
-function find (list, f) {
-  return list.filter(f)[0]
-}
-
-/**
- * Deep copy the given object considering circular structure.
- * This function caches all nested objects and its copies.
- * If it detects circular structure, use cached copy to avoid infinite loop.
- *
- * @param {*} obj
- * @param {Array<Object>} cache
- * @return {*}
- */
-function deepCopy (obj, cache = []) {
-  // just return if obj is immutable value
-  if (obj === null || typeof obj !== 'object') {
-    return obj
-  }
-
-  // if obj is hit, it is in circular structure
-  let hit = find(cache, c => c.original === obj)
-  if (hit) {
-    return hit.copy
-  }
-
-  let copy = Array.isArray(obj) ? [] : {}
-  // put the copy into cache at first
-  // because we want to refer it in recursive deepCopy
-  cache.push({
-    original: obj,
-    copy
-  })
-
-  Object.keys(obj).forEach(key => {
-    copy[key] = deepCopy(obj[key], cache)
-  })
-
-  return copy
-}
-
-module.exports = { deepCopy, find }
