--- utils/forEachValue/index.js
@@ -1,5 +1,0 @@
-function forEachValue (obj, fn) {
-  Object.keys(obj).forEach(key => fn(obj[key], key))
-}
-
-module.exports = { forEachValue }
