--- utils/isPromise/index.js
@@ -1,5 +1,0 @@
-function isPromise (val) {
-  return val && typeof val.then === 'function'
-}
-
-module.exports = { isPromise }
