--- utils/isPromise/index.test.js
@@ -1,9 +1,0 @@
-let { isPromise } = require('.')
-
-it('isPromise', () => {
-  let promise = new Promise(() => {}, () => {})
-  let func = () => {}
-  expect(isPromise(1)).toBe(false)
-  expect(isPromise(promise)).toBe(true)
-  expect(isPromise(func)).toBe(false)
-})
